Added sql products schema (#30)
* Update gradle.yml

* added sql schema

* working sql schema

* style update

* unique constraint added

---------

+++ src/main/java/pl/edu/pw/mini/ingreedio/api/model/ProductCategories.java
@@ -1,7 +1,7 @@
 package pl.edu.pw.mini.ingreedio.api.model;
 
 public class ProductCategories {
-    private Integer id;
-    private Integer productId;
+    private Long id;
+    private Long productId;
     private Category category;
 }
